refactor(CodeEditor): replace createTooltipsExtension with useTooltipExtension and remove unused code
